Add ability to disable Javascript globally
