+++ src/components/preferences/section-privacy/index.js
@@ -14,34 +14,15 @@ import Switch from '@material-ui/core/Switch';
 import ChevronRightIcon from '@material-ui/icons/ChevronRight';
 
 import connectComponent from '../../../helpers/connect-component';
-import checkLicense from '../../../helpers/check-license';
 import getStaticGlobal from '../../../helpers/get-static-global';
 
 import {
   enqueueRequestRestartSnackbar,
   requestClearBrowsingData,
-  requestOpenInBrowser,
   requestSetPreference,
 } from '../../../senders';
 
-const styles = () => ({
-  link: {
-    cursor: 'pointer',
-    fontWeight: 500,
-    outline: 'none',
-    '&:hover': {
-      textDecoration: 'underline',
-    },
-    '&:focus': {
-      textDecoration: 'underline',
-    },
-  },
-});
-
 const SectionBrowsing = ({
-  blockAds,
-  classes,
-  ignoreCertificateErrors,
   rememberLastPageVisited,
   shareWorkspaceBrowsingData,
 }) => {
@@ -52,27 +33,6 @@ const SectionBrowsing = ({
       <ListItem button onClick={requestClearBrowsingData}>
         <ListItemText primary="Clear browsing data" secondary="Clear cookies, cache, and more." />
         <ChevronRightIcon color="action" />
-      </ListItem>
-      <Divider />
-      <ListItem>
-        <ListItemText
-          primary="Block ads &amp; trackers"
-        />
-        <ListItemSecondaryAction>
-          <Switch
-            edge="end"
-            color="primary"
-            checked={blockAds}
-            onChange={(e) => {
-              if (!checkLicense()) {
-                return;
-              }
-
-              requestSetPreference('blockAds', e.target.checked);
-              enqueueRequestRestartSnackbar();
-            }}
-          />
-        </ListItemSecondaryAction>
       </ListItem>
       {appJson.url && (
       <>
@@ -108,56 +68,16 @@ const SectionBrowsing = ({
           />
         </ListItemSecondaryAction>
       </ListItem>
-      <Divider />
-      <ListItem>
-        <ListItemText
-          primary="Ignore certificate errors"
-          secondary={(
-            <>
-              <span>Not recommended. </span>
-              <span
-                role="link"
-                tabIndex={0}
-                className={classes.link}
-                onClick={() => requestOpenInBrowser('https://groups.google.com/a/chromium.org/d/msg/security-dev/mB2KJv_mMzM/ddMteO9RjXEJ')}
-                onKeyDown={(e) => {
-                  if (e.key !== 'Enter') return;
-                  requestOpenInBrowser('https://groups.google.com/a/chromium.org/d/msg/security-dev/mB2KJv_mMzM/ddMteO9RjXEJ');
-                }}
-              >
-                Learn more
-              </span>
-              .
-            </>
-            )}
-        />
-        <ListItemSecondaryAction>
-          <Switch
-            edge="end"
-            color="primary"
-            checked={ignoreCertificateErrors}
-            onChange={(e) => {
-              requestSetPreference('ignoreCertificateErrors', e.target.checked);
-              enqueueRequestRestartSnackbar();
-            }}
-          />
-        </ListItemSecondaryAction>
-      </ListItem>
     </List>
   );
 };
 
 SectionBrowsing.propTypes = {
-  blockAds: PropTypes.bool.isRequired,
-  classes: PropTypes.object.isRequired,
-  ignoreCertificateErrors: PropTypes.bool.isRequired,
   rememberLastPageVisited: PropTypes.bool.isRequired,
   shareWorkspaceBrowsingData: PropTypes.bool.isRequired,
 };
 
 const mapStateToProps = (state) => ({
-  blockAds: state.preferences.blockAds,
-  ignoreCertificateErrors: state.preferences.ignoreCertificateErrors,
   rememberLastPageVisited: state.preferences.rememberLastPageVisited,
   shareWorkspaceBrowsingData: state.preferences.shareWorkspaceBrowsingData,
 });
@@ -165,6 +85,4 @@ const mapStateToProps = (state) => ({
 export default connectComponent(
   SectionBrowsing,
   mapStateToProps,
-  null,
-  styles,
 );
